move raygun into recovery top level
